refactor(player-portal): decompose Progression — split logic out into hook + pure modules
Progression.tsx had grown into a ~1140-line file mixing slot encoding,
flag (de)serialisation, the pick-state machine (3 pick kinds × 2
current/future timing branches × commit/clear = 12 paths of optimistic
update + Foundry write + rollback), hydration from existing items,
auto-apply on level-up, and ~700 lines of JSX. Splitting into a new
`tabs/progression/` directory:

- `slot.ts` — pure slot taxonomy: `SlotType`, `SlotKey`, `slotKey` /
  `parseSlotKey`, `featSlotLocationFor` / `parseFeatLocation`,
  `buildLevelSlotMap`, `groupFeaturesByLevel`, `ABILITY_BOOST_LEVELS`.
- `picks.ts` — `Pick` discriminated union + Foundry flag
  serialisation: `buildProgressionPicksFlags`, `parsePersistedPick`.
- `useProgressionPicks.ts` — the reducer hook that owns the picks Map,
  runs hydration + auto-apply effects, and implements commitPick /
  clearPick with all the optimistic-update + Foundry-write + rollback
  orchestration.

Progression.tsx is now ~730 lines of mostly JSX + the local UI state
(`pickerTarget`) and the picker mounting logic. The doc-prefetch
effect, FeatureChip popover, and per-pick rendering helpers stay in
the component file because they're presentation-layer concerns.

Behaviour is unchanged: `Progression.test.tsx` (which moved into the
new directory) passes against the refactored implementation without
modification.
