+++ apps/player-portal/src/features/characters/sheet/tabs/progression/Progression.tsx
@@ -1,11 +1,9 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { api } from '@/features/characters/api';
-import { ABILITY_KEYS, isClassItem, isFeatItem } from '@/features/characters/types';
+import { isClassItem, isFeatItem } from '@/features/characters/types';
 import type {
   AbilityKey,
   ClassFeatureEntry,
-  ClassItem,
   CompendiumDocument,
   CompendiumMatch,
   CompendiumSearchOptions,
@@ -29,14 +27,9 @@ import {
   pickVerticalSlot,
   clampPopoverLeft,
 } from '@/shared/lib/popover-positioning';
-
-// All picks share the same map keyed by `${level}:${slot}`. The value
-// is a discriminated union so each slot kind can store the shape it
-// needs to reconstruct its chip / later feed into the scratch-actor.
-type Pick =
-  | { kind: 'feat'; match: CompendiumMatch; actorItemId?: string }
-  | { kind: 'skill-increase'; skill: string; newRank: ProficiencyRank }
-  | { kind: 'ability-boosts'; abilities: AbilityKey[] };
+import { buildLevelSlotMap, groupFeaturesByLevel, slotKey, type SlotKey, type SlotType } from './slot';
+import type { Pick } from './picks';
+import { useProgressionPicks } from './useProgressionPicks';
 
 interface Props {
   actorId: string;
@@ -44,25 +37,13 @@ interface Props {
   items: PreparedActorItem[];
   characterContext: CharacterContext;
   onActorChanged: () => void;
-  /** Deserialized from `actor.flags['player-portal']['progression-picks']`.
+  /** Deserialised from `actor.flags['player-portal']['progression-picks']`.
    *  Hydrates skill-increase and ability-boost picks across page refreshes. */
   persistedPicks?: Record<string, unknown>;
 }
 
-// pf2e ability boosts happen at these fixed levels (4 boosts each). This
-// is the only piece of the progression that isn't encoded on the class
-// item itself (every other slot lives in `class.system.*FeatLevels`).
-// See pf2e Core Rulebook "Advancing Your Character" (p.32).
-const ABILITY_BOOST_LEVELS: readonly number[] = [5, 10, 15, 20];
-
 // Levels every character can reach (pf2e core: 1-20).
 const LEVELS: readonly number[] = Array.from({ length: 20 }, (_, i) => i + 1);
-
-// Slot-key for the selection map. One level can open several slot types,
-// and each could eventually have its own pick (two class feats at L12,
-// a class feat + skill feat at L2, etc.), so we key by `${level}:${slot}`.
-type SlotKey = string;
-const slotKey = (level: number, slot: SlotType): SlotKey => `${level.toString()}:${slot}`;
 
 // Progression tab — vertical timeline of levels 1-20, each row showing
 // the auto-granted class features at that level plus chips for every
@@ -77,107 +58,29 @@ const slotKey = (level: number, slot: SlotType): SlotKey => `${level.toString()}
 // Class-feat slots are clickable — they open a CompendiumPicker scoped
 // to the character's class trait and capped at the slot's level (with
 // the creator's prereq + source-filter behavior layered in via
-// useCreatorPickerProps). Picks for current/past levels write to Foundry immediately;
-// picks for future levels are stored in flags only and auto-applied when
-// the character reaches that level.
-export function Progression({ actorId, characterLevel, items, characterContext, onActorChanged, persistedPicks }: Props): React.ReactElement {
+// useCreatorPickerProps). Picks for current/past levels write to Foundry
+// immediately; picks for future levels are stored in flags only and
+// auto-applied when the character reaches that level. The pick state
+// machine + Foundry write orchestration lives in useProgressionPicks;
+// this component is the timeline rendering + UI state (which slot the
+// picker is currently targeting).
+export function Progression({
+  actorId,
+  characterLevel,
+  items,
+  characterContext,
+  onActorChanged,
+  persistedPicks,
+}: Props): React.ReactElement {
   const classItem = items.find(isClassItem);
-  const [picks, setPicks] = useState<Map<SlotKey, Pick>>(new Map());
   const [pickerTarget, setPickerTarget] = useState<{ level: number; slot: SlotType } | null>(null);
-  // Tracks the previous characterLevel so the hydration effect can detect
-  // level-ups and auto-apply picks that just became reachable.
-  const prevCharacterLevelRef = useRef(characterLevel);
-
-  // Hydrate the picks Map from embedded feat items' `system.location`
-  // strings. pf2e tags feats added via its own "Add to Slot" flow
-  // with `<category>-<level>` (e.g. "ancestry-1", "class-2"); the
-  // character-creator wizard sets the same string when it
-  // piecemeal-adds L1 feats. This seeds the chips so the progression
-  // timeline reflects what's already on the actor. Local edits from
-  // within this tab continue to mutate the same Map.
-  useEffect(() => {
-    const hydrated = new Map<SlotKey, Pick>();
-    for (const item of items) {
-      if (item.type !== 'feat') continue;
-      const rawLocation = (item.system as { location?: unknown } | null)?.location;
-      if (typeof rawLocation !== 'string' || rawLocation.length === 0) continue;
-      const parsed = parseFeatLocation(rawLocation);
-      if (parsed === null) continue;
-      hydrated.set(slotKey(parsed.level, parsed.slot), {
-        kind: 'feat',
-        actorItemId: item.id,
-        match: {
-          packId: '',
-          packLabel: '',
-          documentId: item.id,
-          // Use the actor-local item id as a stand-in uuid: the
-          // PreparedActor payload drops `flags.core.sourceId` so we
-          // don't know the compendium origin. Hover previews won't
-          // resolve, but the chip shows name + img correctly.
-          uuid: item.id,
-          name: item.name,
-          type: item.type,
-          img: item.img,
-        },
-      });
-    }
-    // Hydrate non-feat picks from the Foundry actor flag so they survive
-    // a full page refresh. Feat picks are always re-derived from items
-    // (system.location is authoritative); flags are only for the kinds
-    // that Foundry doesn't encode per-slot in the prepared payload.
-    if (persistedPicks !== undefined) {
-      for (const [key, raw] of Object.entries(persistedPicks)) {
-        if (hydrated.has(key)) continue; // feat from items takes precedence
-        const pick = parsePersistedPick(raw);
-        if (pick !== null) hydrated.set(key, pick);
-      }
-    }
-    // eslint-disable-next-line react-hooks/set-state-in-effect
-    setPicks((prev) => {
-      const next = new Map(hydrated);
-      // In-memory fallback: preserve any non-feat picks not yet reflected in
-      // flags (the optimistic window between commitPick and onActorChanged
-      // completing its /prepared reload).
-      for (const [key, pick] of prev) {
-        if (pick.kind !== 'feat' && !next.has(key)) next.set(key, pick);
-      }
-      return next;
-    });
-    // Auto-apply: when the character levels up, fire the deferred system
-    // writes for any non-feat picks that just became reachable (they were
-    // stored in flags while the level was still future).
-    const prevLevel = prevCharacterLevelRef.current;
-    prevCharacterLevelRef.current = characterLevel;
-    if (characterLevel > prevLevel) {
-      for (const [key, pick] of hydrated) {
-        if (pick.kind === 'feat') continue;
-        const parsed = parseSlotKey(key);
-        if (!parsed) continue;
-        const { level } = parsed;
-        if (level <= prevLevel || level > characterLevel) continue;
-        if (pick.kind === 'skill-increase') {
-          void api
-            .updateActor(actorId, {
-              system: { skills: { [pick.skill]: { rank: pick.newRank } } },
-              flags: buildProgressionPicksFlags(hydrated),
-            })
-            .then(() => { onActorChanged(); })
-            .catch((err: unknown) => { console.warn('Failed to auto-apply skill increase', err); });
-        } else if (pick.kind === 'ability-boosts') {
-          void api
-            .updateActor(actorId, {
-              system: { build: { attributes: { boosts: { [level]: pick.abilities } } } },
-              flags: buildProgressionPicksFlags(hydrated),
-            })
-            .then(() => { onActorChanged(); })
-            .catch((err: unknown) => { console.warn('Failed to auto-apply ability boosts', err); });
-        }
-      }
-    }
-    // Re-run when items, persisted flag data, or character level changes —
-    // all three arrive together via the actor refetch triggered by onActorChanged.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [items, persistedPicks, characterLevel]);
+  const { picks, commitPick, clearPick } = useProgressionPicks({
+    actorId,
+    items,
+    characterLevel,
+    ...(persistedPicks !== undefined ? { persistedPicks } : {}),
+    onActorChanged,
+  });
 
   // Prefetched full documents for every class feature on this class,
   // indexed by uuid. A ref-backed cache survives React 18 StrictMode's
@@ -223,10 +126,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
   // are actor-local placeholders, not compendium UUIDs) against the
   // item on the character directly. Fresh picks out of the picker keep
   // their real compendium uuid and hit the API as normal.
-  //
-  // Declared before the `!classItem` early return so React can call
-  // the hooks in the same order on every render.
-  const feathItemById = useMemo(() => {
+  const featItemById = useMemo(() => {
     const map = new Map<string, PreparedActorItem>();
     for (const item of items) {
       if (isFeatItem(item)) map.set(item.id, item);
@@ -235,7 +135,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
   }, [items]);
   const resolveLocalPickDoc = useCallback(
     (uuid: string): CompendiumDocument | undefined => {
-      const item = feathItemById.get(uuid);
+      const item = featItemById.get(uuid);
       if (!item) return undefined;
       return {
         id: item.id,
@@ -246,7 +146,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
         system: item.system,
       };
     },
-    [feathItemById],
+    [featItemById],
   );
   const pickedHover = useUuidHover({ resolveLocal: resolveLocalPickDoc });
 
@@ -276,211 +176,14 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
   const closePicker = (): void => {
     setPickerTarget(null);
   };
-  const commitPick = (pick: Pick): void => {
+  // Bridge from picker callbacks → the hook's state machine. The picker
+  // already knows the slot it's targeting (via `pickerTarget`), so the
+  // component owns the "which slot are we picking for" UI state and the
+  // hook owns the "what does committing mean for this kind" logic.
+  const onPick = (pick: Pick): void => {
     if (!pickerTarget) return;
-    const level = pickerTarget.level;
-    const slot = pickerTarget.slot;
-    const key = slotKey(level, slot);
-    const existingPick = picks.get(key);
-
-    // Compute post-pick state synchronously so flag serialization reflects
-    // the new pick without waiting for React to flush the setState.
-    const newPicksForFlag = new Map(picks);
-    newPicksForFlag.set(key, pick);
-
-    // Optimistic local update — closes the picker immediately.
-    setPicks((prev) => {
-      const next = new Map(prev);
-      next.set(key, pick);
-      return next;
-    });
+    commitPick(pickerTarget.level, pickerTarget.slot, pick);
     setPickerTarget(null);
-
-    const isFutureLevel = level > characterLevel;
-
-    const rollback = (): void => {
-      setPicks((prev) => {
-        const next = new Map(prev);
-        if (existingPick !== undefined) next.set(key, existingPick);
-        else next.delete(key);
-        return next;
-      });
-    };
-
-    if (pick.kind === 'feat') {
-      if (isFutureLevel) {
-        // Future feats are planning-only: keep in local state, no API call.
-        // They'll be applied via addItemFromCompendium when the user reaches
-        // this level and explicitly picks in the now-current slot.
-        return;
-      }
-      const location = featSlotLocationFor(slot, level);
-      void api
-        .addItemFromCompendium(actorId, {
-          packId: pick.match.packId,
-          itemId: pick.match.documentId,
-          ...(location !== null ? { systemOverrides: { location } } : {}),
-        })
-        .then((ref) => {
-          // Attach the actor-local item id so subsequent clears can delete it.
-          setPicks((prev) => {
-            const current = prev.get(key);
-            if (!current || current.kind !== 'feat') return prev;
-            const next = new Map(prev);
-            next.set(key, { ...current, actorItemId: ref.id });
-            return next;
-          });
-          // Remove whichever item was filling this slot before, if any.
-          if (existingPick?.kind === 'feat' && existingPick.actorItemId !== undefined) {
-            void api.deleteActorItem(actorId, existingPick.actorItemId).catch((err: unknown) => {
-              console.warn('Failed to clean up replaced feat item', err);
-            });
-          }
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to persist feat pick', err);
-          rollback();
-        });
-    } else if (pick.kind === 'skill-increase') {
-      if (isFutureLevel) {
-        // Future level: store the planned pick in flags only; the system write
-        // fires automatically when the character reaches this level.
-        void api
-          .updateActor(actorId, { flags: buildProgressionPicksFlags(newPicksForFlag) })
-          .catch((err: unknown) => {
-            console.warn('Failed to save planned skill increase', err);
-            rollback();
-          });
-        return;
-      }
-      void api
-        .updateActor(actorId, {
-          system: { skills: { [pick.skill]: { rank: pick.newRank } } },
-          flags: buildProgressionPicksFlags(newPicksForFlag),
-        })
-        .then(() => {
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to persist skill increase', err);
-          rollback();
-        });
-    } else if (pick.kind === 'ability-boosts') {
-      if (isFutureLevel) {
-        // Future level: store the planned pick in flags only.
-        void api
-          .updateActor(actorId, { flags: buildProgressionPicksFlags(newPicksForFlag) })
-          .catch((err: unknown) => {
-            console.warn('Failed to save planned ability boosts', err);
-            rollback();
-          });
-        return;
-      }
-      void api
-        .updateActor(actorId, {
-          system: { build: { attributes: { boosts: { [level]: pick.abilities } } } },
-          flags: buildProgressionPicksFlags(newPicksForFlag),
-        })
-        .then(() => {
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to persist ability boosts', err);
-          rollback();
-        });
-    }
-  };
-  const clearPick = (level: number, slot: SlotType): void => {
-    const key = slotKey(level, slot);
-    const existingPick = picks.get(key);
-
-    // Compute post-clear state for flag serialization.
-    const newPicksForFlag = new Map(picks);
-    newPicksForFlag.delete(key);
-
-    // Optimistic removal.
-    setPicks((prev) => {
-      const next = new Map(prev);
-      next.delete(key);
-      return next;
-    });
-
-    if (existingPick === undefined) return;
-
-    const isFutureLevel = level > characterLevel;
-
-    const undoClear = (): void => {
-      setPicks((prev) => {
-        const next = new Map(prev);
-        next.set(key, existingPick);
-        return next;
-      });
-    };
-
-    if (existingPick.kind === 'feat') {
-      if (isFutureLevel) {
-        // Future feat was never added to the actor — nothing to undo in Foundry.
-        return;
-      }
-      if (existingPick.actorItemId === undefined) return;
-      void api
-        .deleteActorItem(actorId, existingPick.actorItemId)
-        .then(() => {
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to delete feat item', err);
-          undoClear();
-        });
-    } else if (existingPick.kind === 'skill-increase') {
-      if (isFutureLevel) {
-        // Planned pick was never written to system — only remove from flags.
-        void api
-          .updateActor(actorId, { flags: buildProgressionPicksFlags(newPicksForFlag) })
-          .catch((err: unknown) => {
-            console.warn('Failed to remove planned skill increase from flags', err);
-            undoClear();
-          });
-        return;
-      }
-      const prevRank = (existingPick.newRank - 1) as ProficiencyRank;
-      void api
-        .updateActor(actorId, {
-          system: { skills: { [existingPick.skill]: { rank: prevRank } } },
-          flags: buildProgressionPicksFlags(newPicksForFlag),
-        })
-        .then(() => {
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to revert skill increase', err);
-          undoClear();
-        });
-    } else if (existingPick.kind === 'ability-boosts') {
-      if (isFutureLevel) {
-        // Planned pick was never written to system — only remove from flags.
-        void api
-          .updateActor(actorId, { flags: buildProgressionPicksFlags(newPicksForFlag) })
-          .catch((err: unknown) => {
-            console.warn('Failed to remove planned ability boosts from flags', err);
-            undoClear();
-          });
-        return;
-      }
-      void api
-        .updateActor(actorId, {
-          system: { build: { attributes: { boosts: { [level]: [] } } } },
-          flags: buildProgressionPicksFlags(newPicksForFlag),
-        })
-        .then(() => {
-          onActorChanged();
-        })
-        .catch((err: unknown) => {
-          console.warn('Failed to clear ability boosts', err);
-          undoClear();
-        });
-    }
   };
 
   const pickerFilters = pickerTarget
@@ -521,7 +224,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
           filters={pickerFilters}
           characterContext={characterContext}
           onPick={(match): void => {
-            commitPick({ kind: 'feat', match });
+            onPick({ kind: 'feat', match });
           }}
           onClose={closePicker}
         />
@@ -531,7 +234,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
           level={pickerTarget.level}
           characterContext={characterContext}
           onPick={(skill, newRank): void => {
-            commitPick({ kind: 'skill-increase', skill, newRank });
+            onPick({ kind: 'skill-increase', skill, newRank });
           }}
           onClose={closePicker}
         />
@@ -543,7 +246,7 @@ export function Progression({ actorId, characterLevel, items, characterContext, 
             level: pickerTarget.level,
             characterContext,
             onPick: (abilities: AbilityKey[]): void => {
-              commitPick({ kind: 'ability-boosts', abilities });
+              onPick({ kind: 'ability-boosts', abilities });
             },
             onClose: closePicker,
             ...(seed ? { initialSelection: seed } : {}),
@@ -819,8 +522,6 @@ function FeatureChip({
 
 // ─── Slot chips ────────────────────────────────────────────────────────
 
-type SlotType = 'class-feat' | 'ancestry-feat' | 'skill-feat' | 'general-feat' | 'skill-increase' | 'ability-boosts';
-
 const SLOT_LABEL: Record<SlotType, string> = {
   'class-feat': 'Class Feat',
   'ancestry-feat': 'Ancestry Feat',
@@ -991,118 +692,9 @@ function capitaliseSkillSlug(s: string): string {
     .join(' ');
 }
 
-// ─── Helpers ───────────────────────────────────────────────────────────
-
-// Map a Progression slot type to the pf2e `<category>-<level>` location
-// string written into `feat.system.location`. Returns null for slot types
-// that don't use a location tag (none currently, but guards future slots).
-const FEAT_SLOT_LOCATION_PREFIX: Partial<Record<SlotType, string>> = {
-  'class-feat': 'class',
-  'ancestry-feat': 'ancestry',
-  'skill-feat': 'skill',
-  'general-feat': 'general',
-};
-
-function featSlotLocationFor(slot: SlotType, level: number): string | null {
-  const prefix = FEAT_SLOT_LOCATION_PREFIX[slot];
-  return prefix !== undefined ? `${prefix}-${level.toString()}` : null;
-}
-
-// Reverse of slotKey — used by the auto-apply logic to recover the level
-// and slot type from a stored Map key without keeping them separately.
-function parseSlotKey(key: SlotKey): { level: number; slot: SlotType } | null {
-  const sep = key.indexOf(':');
-  if (sep === -1) return null;
-  const level = Number(key.slice(0, sep));
-  if (!Number.isFinite(level)) return null;
-  return { level, slot: key.slice(sep + 1) as SlotType };
-}
-
-// Serialise all non-feat picks into the Foundry actor flag shape. Written to
-// flags['player-portal']['progression-picks'] so picks survive page refreshes.
-function buildProgressionPicksFlags(picks: Map<SlotKey, Pick>): Record<string, Record<string, unknown>> {
-  const blob: Record<string, unknown> = {};
-  for (const [key, pick] of picks) {
-    if (pick.kind !== 'feat') blob[key] = pick;
-  }
-  return { 'player-portal': { 'progression-picks': blob } };
-}
-
-// Deserialise one entry from the stored flag blob back into a typed Pick.
-// Returns null for any shape that doesn't match a known non-feat kind.
-function parsePersistedPick(raw: unknown): Exclude<Pick, { kind: 'feat' }> | null {
-  if (!raw || typeof raw !== 'object') return null;
-  const obj = raw as Record<string, unknown>;
-  if (obj.kind === 'skill-increase') {
-    if (typeof obj.skill !== 'string') return null;
-    const rank = obj.newRank;
-    if (typeof rank !== 'number' || rank < 0 || rank > 4) return null;
-    return { kind: 'skill-increase', skill: obj.skill, newRank: rank as ProficiencyRank };
-  }
-  if (obj.kind === 'ability-boosts') {
-    if (!Array.isArray(obj.abilities)) return null;
-    const abilities = (obj.abilities as unknown[]).filter(
-      (a): a is AbilityKey => typeof a === 'string' && (ABILITY_KEYS as readonly string[]).includes(a),
-    );
-    return { kind: 'ability-boosts', abilities };
-  }
-  return null;
-}
-
-function groupFeaturesByLevel(items: ClassItem['system']['items']): Map<number, ClassFeatureEntry[]> {
-  const out = new Map<number, ClassFeatureEntry[]>();
-  for (const entry of Object.values(items)) {
-    const arr = out.get(entry.level) ?? [];
-    arr.push(entry);
-    out.set(entry.level, arr);
-  }
-  for (const [, arr] of out) arr.sort((a, b) => a.name.localeCompare(b.name));
-  return out;
-}
-
-// Parse pf2e's `<category>-<level>` slot location strings into the
-// Progression tab's slot taxonomy. Returns null for anything we
-// don't model (archetype-N, etc. — pf2e archetype feats live
-// outside the stock level chassis).
-function parseFeatLocation(location: string): { slot: SlotType; level: number } | null {
-  const match = /^(ancestry|class|skill|general)-(\d+)$/.exec(location);
-  if (!match) return null;
-  const level = Number(match[2]);
-  if (!Number.isFinite(level) || level < 1) return null;
-  const prefix = match[1] as 'ancestry' | 'class' | 'skill' | 'general';
-  return { slot: `${prefix}-feat`, level };
-}
-
-function buildLevelSlotMap(sys: ClassItem['system']): Map<number, readonly SlotType[]> {
-  // Render order for slots on a given level. Class feats come first because
-  // they're the most character-defining; ability boosts last because they
-  // collapse into a single "4 boosts" chip.
-  const rules: Array<[SlotType, readonly number[]]> = [
-    ['class-feat', sys.classFeatLevels.value],
-    ['ancestry-feat', sys.ancestryFeatLevels.value],
-    ['skill-feat', sys.skillFeatLevels.value],
-    ['general-feat', sys.generalFeatLevels.value],
-    ['skill-increase', sys.skillIncreaseLevels.value],
-    ['ability-boosts', ABILITY_BOOST_LEVELS],
-  ];
-  const out = new Map<number, SlotType[]>();
-  for (const [slot, levels] of rules) {
-    for (const level of levels) {
-      const arr = out.get(level) ?? [];
-      arr.push(slot);
-      out.set(level, arr);
-    }
-  }
-  return out;
-}
-
 // Wraps CompendiumPicker with the creator's prereq + sort + source-filter
 // behavior. Keying on level+slot remounts this on each open so picker
 // state resets per slot.
-// `Pick` is shadowed by the local progression-pick discriminated-union
-// type, so we use the global helper via this alias to work around the
-// scoping collision. (Putting this at module scope captures the
-// built-in Pick before the local declaration.)
 type _CreatorFilters = {
   [K in
     | 'packIds'
